+++ dokan_fuse/src/utils.cpp
@@ -312,7 +312,7 @@ static const struct errentry errtable[] = {
 };
 const int errtable_size = sizeof(errtable) / sizeof(errtable[0]);
 
-extern "C" int ntstatus_error_to_errno(int win_res) {
+extern "C" int ntstatus_error_to_errno(long win_res) {
   if (win_res == 0)
     return 0; // No error
 
@@ -324,7 +324,7 @@ extern "C" int ntstatus_error_to_errno(int win_res) {
   return EINVAL;
 }
 
-extern "C" int errno_to_ntstatus_error(int err) {
+extern "C" long errno_to_ntstatus_error(int err) {
   if (err == 0)
     return 0; // No error
